requirements.txt
langchain-core
langchain-ollama
langchain-groq
aio-pika
python-dotenv
aiohttp